# backend/app/api/routes/seniors.py
from __future__ import annotations

import json
import random
import string
from datetime import datetime, timedelta, timezone

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.orm import Session

from app.core.database import get_db
from app.core.deps import get_current_user
from app.models.senior import Senior
from app.models.session import Session as ConvSession
from app.models.user import FamilyMember, User
from app.schemas.senior import (
    FamilyMemberInfo,
    LastSessionInfo,
    ScheduleInfo,
    SeniorCreate,
    SeniorDetailResponse,
    SeniorResponse,
    SeniorUpdate,
)

router = APIRouter(prefix="/seniors", tags=["seniors"])


@router.post("/", response_model=SeniorResponse, status_code=status.HTTP_201_CREATED)
def create_senior(
    data: SeniorCreate,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    senior = Senior(**data.model_dump())
    db.add(senior)
    db.flush()

    link = FamilyMember(user_id=current_user.id, senior_id=senior.id, role="family")
    db.add(link)
    db.commit()
    db.refresh(senior)
    return senior


@router.get("/", response_model=list[SeniorResponse])
def list_seniors(
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    senior_ids = (
        db.query(FamilyMember.senior_id)
        .filter(FamilyMember.user_id == current_user.id)
        .subquery()
    )
    seniors = db.query(Senior).filter(Senior.id.in_(senior_ids)).all()
    return seniors


@router.get("/{senior_id}", response_model=SeniorDetailResponse)
def get_senior(
    senior_id: int,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    senior = db.query(Senior).filter(Senior.id == senior_id).first()
    if not senior:
        raise HTTPException(status_code=404, detail="Senior introuvable")

    link = (
        db.query(FamilyMember)
        .filter(FamilyMember.user_id == current_user.id, FamilyMember.senior_id == senior_id)
        .first()
    )
    if not link:
        raise HTTPException(status_code=403, detail="Acces non autorise")

    # Session schedule (stored as JSON on the senior)
    schedule = None
    if senior.session_schedule:
        try:
            raw = json.loads(senior.session_schedule)
            schedule = ScheduleInfo(
                days=raw.get("days", []),
                time=raw.get("time", "10:00"),
                duration_minutes=raw.get("duration_minutes", 30),
            )
        except (json.JSONDecodeError, TypeError, AttributeError):
            schedule = None

    # Family members, resolved to the linked user's name/email
    members = [
        FamilyMemberInfo(
            id=fm.id,
            name=f"{fm.user.first_name} {fm.user.last_name}".strip() if fm.user else "",
            email=fm.user.email if fm.user else "",
            role=fm.role,
        )
        for fm in senior.family_members
    ]

    # Most recent session for the dashboard's "last session" card
    last = (
        db.query(ConvSession)
        .filter(ConvSession.senior_id == senior_id)
        .order_by(ConvSession.started_at.desc())
        .first()
    )
    last_session = (
        LastSessionInfo(id=last.id, date=last.started_at, summary=last.summary) if last else None
    )

    base = SeniorResponse.model_validate(senior)
    return SeniorDetailResponse(
        **base.model_dump(),
        schedule=schedule,
        family_members=members,
        last_session=last_session,
    )


@router.put("/{senior_id}", response_model=SeniorResponse)
def update_senior(
    senior_id: int,
    data: SeniorUpdate,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    senior = db.query(Senior).filter(Senior.id == senior_id).first()
    if not senior:
        raise HTTPException(status_code=404, detail="Senior introuvable")

    link = (
        db.query(FamilyMember)
        .filter(FamilyMember.user_id == current_user.id, FamilyMember.senior_id == senior_id)
        .first()
    )
    if not link:
        raise HTTPException(status_code=403, detail="Acces non autorise")

    for field, value in data.model_dump(exclude_unset=True).items():
        setattr(senior, field, value)

    db.commit()
    db.refresh(senior)
    return senior


@router.post("/{senior_id}/pairing-code")
def generate_pairing_code(
    senior_id: int,
    db: Session = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    senior = db.query(Senior).filter(Senior.id == senior_id).first()
    if not senior:
        raise HTTPException(status_code=404, detail="Senior introuvable")

    link = (
        db.query(FamilyMember)
        .filter(FamilyMember.user_id == current_user.id, FamilyMember.senior_id == senior_id)
        .first()
    )
    if not link:
        raise HTTPException(status_code=403, detail="Acces non autorise")

    # Generate a random 6-digit pairing code + 4-digit settings PIN
    code = "".join(random.choices(string.digits, k=6))
    expires_at = datetime.now(timezone.utc) + timedelta(hours=24)

    # Store in the senior's preferences JSON field
    prefs: dict = {}
    if senior.preferences:
        try:
            prefs = json.loads(senior.preferences)
        except (json.JSONDecodeError, TypeError):
            prefs = {}

    prefs["pairing_code"] = code
    prefs["pairing_expires"] = expires_at.isoformat()

    # Generate settings PIN if not already set
    if not prefs.get("settings_pin"):
        prefs["settings_pin"] = "".join(random.choices(string.digits, k=4))

    senior.preferences = json.dumps(prefs)
    db.commit()

    return {
        "code": code,
        "settings_pin": prefs["settings_pin"],
        "expires_in": 86400,
    }